# tsoa/src/notes/noteController.js
"use strict";
var __extends = (this && this.__extends) || (function () {
    var extendStatics = function (d, b) {
        extendStatics = Object.setPrototypeOf ||
            ({ __proto__: [] } instanceof Array && function (d, b) { d.__proto__ = b; }) ||
            function (d, b) { for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p]; };
        return extendStatics(d, b);
    };
    return function (d, b) {
        extendStatics(d, b);
        function __() { this.constructor = d; }
        d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
    };
})();
var __decorate = (this && this.__decorate) || function (decorators, target, key, desc) {
    var c = arguments.length, r = c < 3 ? target : desc === null ? desc = Object.getOwnPropertyDescriptor(target, key) : desc, d;
    if (typeof Reflect === "object" && typeof Reflect.decorate === "function") r = Reflect.decorate(decorators, target, key, desc);
    else for (var i = decorators.length - 1; i >= 0; i--) if (d = decorators[i]) r = (c < 3 ? d(r) : c > 3 ? d(target, key, r) : d(target, key)) || r;
    return c > 3 && r && Object.defineProperty(target, key, r), r;
};
var __param = (this && this.__param) || function (paramIndex, decorator) {
    return function (target, key) { decorator(target, key, paramIndex); }
};
var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
    return new (P || (P = Promise))(function (resolve, reject) {
        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
        step((generator = generator.apply(thisArg, _arguments || [])).next());
    });
};
var __generator = (this && this.__generator) || function (thisArg, body) {
    var _ = { label: 0, sent: function() { if (t[0] & 1) throw t[1]; return t[1]; }, trys: [], ops: [] }, f, y, t, g;
    return g = { next: verb(0), "throw": verb(1), "return": verb(2) }, typeof Symbol === "function" && (g[Symbol.iterator] = function() { return this; }), g;
    function verb(n) { return function (v) { return step([n, v]); }; }
    function step(op) {
        if (f) throw new TypeError("Generator is already executing.");
        while (_) try {
            if (f = 1, y && (t = op[0] & 2 ? y["return"] : op[0] ? y["throw"] || ((t = y["return"]) && t.call(y), 0) : y.next) && !(t = t.call(y, op[1])).done) return t;
            if (y = 0, t) op = [op[0] & 2, t.value];
            switch (op[0]) {
                case 0: case 1: t = op; break;
                case 4: _.label++; return { value: op[1], done: false };
                case 5: _.label++; y = op[1]; op = [0]; continue;
                case 7: op = _.ops.pop(); _.trys.pop(); continue;
                default:
                    if (!(t = _.trys, t = t.length > 0 && t[t.length - 1]) && (op[0] === 6 || op[0] === 2)) { _ = 0; continue; }
                    if (op[0] === 3 && (!t || (op[1] > t[0] && op[1] < t[3]))) { _.label = op[1]; break; }
                    if (op[0] === 6 && _.label < t[1]) { _.label = t[1]; t = op; break; }
                    if (t && _.label < t[2]) { _.label = t[2]; _.ops.push(op); break; }
                    if (t[2]) _.ops.pop();
                    _.trys.pop(); continue;
            }
            op = body.call(thisArg, _);
        } catch (e) { op = [6, e]; y = 0; } finally { f = t = 0; }
        if (op[0] & 5) throw op[1]; return { value: op[0] ? op[1] : void 0, done: true };
    }
};
Object.defineProperty(exports, "__esModule", { value: true });
exports.NotesController = void 0;
var tsoa_1 = require("tsoa");
var notesService_1 = require("./notesService");
//ko želiš vse zagnati, moraš še enkrat znova zagnati ukaz yarn run tsoa routes
// in nato še yarn run tsc --outDir build --experimentalDecorators
//ko zaženeš drugi klic, se ponovno zgenerirajo vse js datoteke, potrebne za delovanje (prek tsoa)
var NotesController = /** @class */ (function (_super) {
    __extends(NotesController, _super);
    function NotesController() {
        return _super !== null && _super.apply(this, arguments) || this;
    }
    NotesController.prototype.getNotes = function () {
        return __awaiter(this, void 0, void 0, function () {
            return __generator(this, function (_a) {
                return [2 /*return*/, new notesService_1.NotesService().getNotes()];
            });
        });
    };
    NotesController.prototype.findOneNote = function (id) {
        return __awaiter(this, void 0, void 0, function () {
            return __generator(this, function (_a) {
                return [2 /*return*/, new notesService_1.NotesService().findOneNote(id)];
            });
        });
    };
    NotesController.prototype.insertNote = function (requestBody) {
        return __awaiter(this, void 0, void 0, function () {
            return __generator(this, function (_a) {
                return [2 /*return*/, new notesService_1.NotesService().insertNote(requestBody)];
            });
        });
    };
    NotesController.prototype.updateNote = function (requestBody) {
        return __awaiter(this, void 0, void 0, function () {
            return __generator(this, function (_a) {
                new notesService_1.NotesService().updateNote(requestBody);
                return [2 /*return*/];
            });
        });
    };
    NotesController.prototype.deleteNote = function (id) {
        return __awaiter(this, void 0, void 0, function () {
            return __generator(this, function (_a) {
                new notesService_1.NotesService().deleteNote(id);
                return [2 /*return*/];
            });
        });
    };
    __decorate([
        tsoa_1.Get("list")
    ], NotesController.prototype, "getNotes", null);
    __decorate([
        tsoa_1.Get("/{id}")
    ], NotesController.prototype, "findOneNote", null);
    __decorate([
        tsoa_1.Post("/"),
        __param(0, tsoa_1.Body())
    ], NotesController.prototype, "insertNote", null);
    __decorate([
        tsoa_1.Post("update"),
        __param(0, tsoa_1.Body())
    ], NotesController.prototype, "updateNote", null);
    __decorate([
        tsoa_1.Delete("delete"),
        __param(0, tsoa_1.Body())
    ], NotesController.prototype, "deleteNote", null);
    NotesController = __decorate([
        tsoa_1.Route("/api/note")
    ], NotesController);
    return NotesController;
}(tsoa_1.Controller));
exports.NotesController = NotesController;